refactor: Icon 컴포넌트 추출
