+++ src/components/shared/index.tsx
@@ -7,3 +7,4 @@ export { default as Select } from './Select/Select';
 export { default as Spinner } from './Spinner/Spinner';
 export { default as Main } from './Main/Main';
 export { default as Heading1 } from './Heading/Heading1';
+export { default as Icon } from './Icon/Icon';
