+++ src/pages/LoginPage/LoginPage.tsx
@@ -2,7 +2,7 @@ import { FormEventHandler, useState, useContext } from 'react';
 import { MdEmail, MdLock } from 'react-icons/md';
 import { Redirect, useHistory } from 'react-router-dom';
 
-import { Box, Button, Input, InputContainer, Heading1 } from '../../components/shared';
+import { Box, Button, Input, InputContainer, Heading1, Icon } from '../../components/shared';
 
 import { UserContext } from '../../contexts/UserContextProvider';
 import { ThemeContext } from '../../contexts/ThemeContextProvider';
@@ -16,7 +16,7 @@ import { ERROR_MESSAGE, SUCCESS_MESSAGE } from '../../constants/messages';
 import useInput from '../../hooks/useInput';
 import apiRequest from '../../request';
 import { PageProps } from '../types';
-import { Icon, SignUpLink, ErrorText, Form } from './LoginPage.style';
+import { SignUpLink, ErrorText, Form } from './LoginPage.style';
 
 const LoginPage = ({ setIsLoading }: PageProps) => {
   const [email, onEmailChange] = useInput('');
